new sortDocumentName function
